refactor: move highlighted cell styles to CellData
