+++ src/pages/tables/generated-table/GeneratedTableCell.js
@@ -1,34 +1,14 @@
-import React, { useState } from 'react'
+import React from 'react'
 import { TableCell } from '@dhis2/ui'
 import PropTypes from 'prop-types'
 import CellData from './CellData'
 import { DATA, TEXT, EMPTY } from '../../../modules/contentTypes'
-import { useTableState } from '../../../context/tableContext'
-
-function getColor(intervals, value) {
-    for (const { lowerBound, color } of intervals) {
-        if (Number(value) >= Number(lowerBound)) return color
-    }
-}
 
 export function GeneratedTableCell({
     cell,
     selectedOrgUnits,
     selectedPeriods,
 }) {
-    const [cellColor, setCellColor] = useState()
-    const table = useTableState()
-
-    function onLoad(value) {
-        if (table.highlightingOn) {
-            const color = getColor(
-                cell.highlightingIntervals || table.highlightingIntervals,
-                value
-            )
-            setCellColor(color)
-        }
-    }
-
     function getCellContents() {
         if (!cell) return null
         switch (cell.contentType) {
@@ -38,7 +18,6 @@ export function GeneratedTableCell({
                         cell={cell}
                         selectedOrgUnits={selectedOrgUnits}
                         selectedPeriods={selectedPeriods}
-                        onLoad={onLoad}
                     />
                 )
             case TEXT:
@@ -49,21 +28,7 @@ export function GeneratedTableCell({
         }
     }
 
-    return (
-        <TableCell>
-            <div>
-                {getCellContents()}
-                <style jsx>{`
-                    div {
-                        display: inline-block;
-                        padding: 0.5rem;
-                        margin: -0.5rem;
-                        background-color: ${cellColor};
-                    }
-                `}</style>
-            </div>
-        </TableCell>
-    )
+    return <TableCell>{getCellContents()}</TableCell>
 }
 
 GeneratedTableCell.propTypes = {
